refactor: organize files into page and component folders
