+++ src/components/Layout/Layout.tsx
@@ -1,4 +1,3 @@
-import './Home.module.css'
 import Header from '@/components/Header/Header.tsx'
 import { Outlet } from 'react-router-dom'
 import Container from '@/components/Container/Container.tsx'
