+++ src/main.tsx
@@ -1,10 +1,10 @@
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import Layout from './Layout.tsx'
+import Layout from '@/components/Layout/Layout.tsx'
 import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom'
-import Home from '@/Home.tsx'
-import LogementDetail from '@/pages/LogementDetail.tsx'
-import NotFound from '@/pages/NotFound.tsx'
+import Home from '@/pages/Home/Home.tsx'
+import LogementDetail from '@/pages/LogementDetail/LogementDetail.tsx'
+import NotFound from '@/pages/NotFound/NotFound.tsx'
 import { logementLoader } from '@/loaders/logement.loader.ts'
 
 const router = createBrowserRouter([
